+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-03T12:01:01.140Z",
+    "2026-04-03T15:25:58.138Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {