+++ packages/orq-rc/RELEASES.md
@@ -8019,3 +8019,13 @@ Based on:
 - [typescript v4.7.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.45 - packages/orq-rc
+
+## 2026-04-03 15:29:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.46 - packages/orq-rc